+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Image tag
 - Target a character using @char(name) in the command tag (e.g. `# command "attribute @char(user) minecraft:scale base set 3"`)
 - Assign custom nbt to characters
+- Render image in dialog `Clara: [img clara/happy] I'm feeling good right now!`
 
 ### Fixes
 - Dialog 2D scale not fixed based on gui scale